Ensure that changes are saved when adjustments mode is closed.
